processor/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

# Core API
fastapi
uvicorn[standard]
python-multipart

# NumPy
numpy<2

# Image & OCR
pillow
opencv-python-headless
pytesseract
pdf2image

# ML / Vision
torch==2.2.2+cpu
torchvision==0.17.2+cpu
torchaudio==2.2.2+cpu
ultralytics
transformers

# Google Vision
google-cloud-vision
google-api-core
google-auth
protobuf<5
python-multipart